university_drf/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.24
Django==3.2.8
djangorestframework==3.12.4
download==0.3.5
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
kombu==5.2.0
libsvmdata==0.3
matplotlib==3.4.3
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2==2.9.1
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
xarray==0.20.1
